pyproject.toml
[tool.poetry]
name = "teste"
version = "0.1.0"
description = ""
authors = ["Haruan"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
mongoengine = "^0.21.0"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
pylint = "^2.6.0"
flake8 = "^3.8.4"
black = "^20.8b1"
jedi = "^0.17.2"
rope = "^0.18.0"
ropevim = "^0.8.0"
mongomock = "^3.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
